Add "node_modules" to .gitignore
We are adding the "node_modules" directory to the .gitignore file in the project. This change is necessary to prevent unnecessary and bulky dependencies from being tracked in version control.
By adding "node_modules" to .gitignore, we ensure that the project's repository remains clean and focused on the essential codebase. Developers can still install the required dependencies using package.json and package-lock.json (or yarn.lock for Yarn users) when setting up the project locally.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>
Co-authored-by: Eduardoespl <eduardoesp3317@gmail.com>
Co-authored-by: Antonio3veces3 <aantonioramirez33@gmail.com>