Create reusable Button component
Create a new Button component to improve code organization and reusability. The Button component is designed to be easily customizable with different text and onPress functionality.

The component receives the following props:
- `text`: A string that represents the button label.
- `onPress`: A function to handle the button press event.

When the Button component is rendered, it displays a TouchableOpacity with the provided text inside. The onPress function is triggered when the button is pressed, allowing for custom actions to be implemented in parent components.

The Button component can now be used throughout the application, promoting consistent design and behavior for buttons. Initially, the onPress prop is set to throw an error to indicate that the functionality is not yet implemented. This will help remind me as developer to update the onPress prop with the appropriate functionality when using the Button in different parts of the app.

Files Modified:
- Button.tsx: The new reusable Button component implementation.
- ProfileClientForm.tsx: Updated to use the Button component for the "Save" button.

This commit is a step towards improving the maintainability and readability of the codebase by encapsulating button logic into a separate component.