Create LocationInput.tsx Component with API Fetch for Autocomplete
Created the LocationInput.tsx component that includes fetching from a specified API URL to manage location autocompletion as the user types into the <TextInput>. The suggestions are displayed in a <FlatList> with logic to close once the user selects an address. The component also works in conjunction with the custom hook useCustomForm.ts.

The structure also anticipates the creation of the LocationInputSchema.ts file within the schema folder, where the interface and types for AutocompleteProps will be defined. This is still pending for future commits, but the necessary planning within the code has been made thanks to an in-depth understanding of the programming requirements.