Add Custom Loading Animation to Assets for MessageModal Component
Created a new animations folder inside the assets directory and added the LoadingAnimation.json file, containing the customized loading animation. This animation was downloaded by me from the LottieAnimations website and tailored for the project. It will be used specifically in the previously created MessageModal component, enhancing the visual feedback provided to users during loading processes.