Update onSubmit Function in useProfileForm Hook to Clear Fields in ProfileClientForm.
Updated the onSubmit function within the useProfileForm hook to clear specific fields in the ProfileClientForm after saving. Calls to setValue were added for each of the fields that need to be cleared, namely name, phone, birthDate, and location. These lines reset the values to an empty string, effectively clearing them, ensuring that the form is ready for a new input or submission.