Utilizing KeyboardAwareScrollView: Enhancing Keyboard Behavior on Screens.
On ProfileClientScreen and ProfileProviderScreen:

I instead of using a simple ScrollView, I decided to use KeyboardAwareScrollView, which is an improved way to handle the keyboard's behavior on the screen.

Moving Container Styles to contentContainerStyle: I moved the styles originally applied to the container (such as alignItems) to the contentContainerStyle property of the KeyboardAwareScrollView. This change prevented an error from occurring.

Enabling Scrolling: I switched the scrollEnabled property to true (or simply removed the line if it was set to false), allowing the content to be scrollable when needed.

Style Adjustments: I modified the styles to match the updated structure and allow everything to fit together correctly. These adjustments helped to create a more user-friendly experience, and the use of KeyboardAwareScrollView provided a superior solution to the previous approach.