Refactor CustomCalendar Component: Extract Schemas and Types.
In an effort to organize and enhance the maintainability of the code, the CustomCalendar component underwent a refactoring process:

Separation of Concerns: Schemas and types related to the CustomCalendar component were extracted and moved to a separate file named CustomCalendarSchema.ts.

Location: This new file was placed within the schema folder.

Also deleted unused components from react-native modules.

Importing: The extracted schemas and types were then imported back into the CustomCalendar component as required.
This change helps to keep the codebase clean and modular, allowing for easier understanding and potential modifications in the future.