Refactored AppointmentList to use the new schema
With the new schema, the AppointmentList component was refactored to use this, and to fill the data in the modal with the data.
Also updated the AppointmentDetailsModal to use the new schema and get the image from the data instead of the props.
Implemented a pressable to trigger the modal opening when the user presses the appointment card.
Showing the details of the appointment and the description, with a button to close the modal.