mplement Firebase Database Getters and Setters for Provider and Client
In this commit, we have implemented the necessary getters and setters
for interacting with our Firebase database in regard to the Provider and
Client entities.

Getters:
Functions that allow us to retrieve Provider and Client data from the
Firebase database.
Setters:
Functions that allow us to update Provider and Client data in the
Firebase database.
By having these dedicated functions for interacting with our database,
we can ensure a more structured approach to data management, which in
turn makes our application more robust and maintainable.

Co-authored-by: DerekAyala <derek.ayala12345@hotmail.com>
Co-authored-by: Eduardoespl <eduardoesp3317@gmail.com>
Co-authored-by: Antonio3veces3 <aantonioramirez33@gmail.com>