mplement UserContext for Managing userType Across the App
The UserContext was implemented to manage and share the userType
information across the entire application. This significant update comes
with various benefits and enhancements:

Creating Context:
Created a context called UserContext to manage and share userType
information across the application.
Defined a UserProvider that wraps the application, allowing components
to access the context.

Using Context in Components:
In WelcomeScreen, used the useUserContext hook to access the setUserType
function from the context and stored the userType when the user chooses
a profile.
In SignInScreen, accessed userType directly from the context using the
useUserContext hook, which allowed utilizing it for navigation
post-login.
In HomeClientScreen, used the useUserContext hook again to access
userType from the context, and used it to display the correct
information based on the user type.

Removing the Need for Navigation Props:
No longer needed to pass userType through navigation properties across
different screens. Instead, stored it in the context and retrieved it as
needed.

Benefits Achieved:
Simplified the data flow in the application, removing the complexity of
passing props between different components.
Achieved cleaner and more efficient data handling by centralizing user
information in one place.
Reduced the possibilities of data loss-related errors between screens
and improved consistency throughout the application.

In summary, successfully implemented a context in the React Native
application to manage and share userType information, which simplified
the workflow and improved the way components access and utilize this
information across different parts of the application